fixed *if no member specified, command will break*

+++ src/commands/Kick.js
@@ -8,24 +8,29 @@ module.exports = {
             const reason = args.slice(1).join(' ');
             const auditLog_QBom = require ("../config.json")
             const auditLog = message.guild.channels.cache.find(channel => channel.id === auditLog_QBom);
-            const kickedEmbed = {
-                color: '#FFEA00',
-                title: `Usuário kickado: ${member.user.tag}`,
-                author: {
-                    name: `${message.author.tag}`,
-                    icon_url: `${message.author.avatarURL()}`,
-                },
-                fields: [
-                    {
-                        name: 'Motivo:',
-                        value: `${reason}.`,
+            if (member != String) {
+                message.reply("o comando está incompleto!");
+            }
+            else {
+                const kickedEmbed = {
+                    color: '#FFEA00',
+                    title: `Usuário kickado: ${member.user.tag}`,
+                    author: {
+                        name: `${message.author.tag}`,
+                        icon_url: `${message.author.avatarURL()}`,
                     },
-                ],
-                timestamp: new Date(),
-            };
-            message.delete();
-            message.guild.member(member).kick(reason);
-            auditLog.send({embed: kickedEmbed});
+                    fields: [
+                        {
+                            name: 'Motivo:',
+                            value: `${reason}.`,
+                        },
+                    ],
+                    timestamp: new Date(),
+                };
+                message.delete();
+                message.guild.member(member).kick(reason);
+                auditLog.send({embed: kickedEmbed});
+            }
         }
         else (message.channel.send(`${message.author}, você não pode usar este comando!`))
     }
